+++ ali/common/src/test/java/com/yanny/ali/platform/TestPlatformHelper.java
@@ -12,6 +12,8 @@ import java.nio.file.Path;
 import java.util.List;
 
 public class TestPlatformHelper implements IPlatformHelper {
+    public static Path CONFIG_DIR = null;
+
     @NotNull
     @Override
     public List<IPlugin> getPlugins() {
@@ -21,7 +23,7 @@ public class TestPlatformHelper implements IPlatformHelper {
     @NotNull
     @Override
     public Path getConfiguration() {
-        return null;
+        return CONFIG_DIR;
     }
 
     @Nullable