+++ awi/common/src/main/java/com/yanny/awi/manager/AwiCommonRegistry.java
@@ -3,12 +3,14 @@ package com.yanny.awi.manager;
 import com.yanny.aci.manager.CoreCommonRegistry;
 import com.yanny.awi.api.ICommonRegistry;
 import com.yanny.awi.api.ICommonUtils;
+import com.yanny.awi.configuration.AwiConfig;
+import com.yanny.awi.configuration.ConfigUtils;
 import org.jetbrains.annotations.NotNull;
 
-public class AwiCommonRegistry extends CoreCommonRegistry<Object> implements ICommonRegistry, ICommonUtils {
+public class AwiCommonRegistry extends CoreCommonRegistry<AwiConfig> implements ICommonRegistry, ICommonUtils {
     @NotNull
     @Override
-    protected Object loadConfiguration() {
-        return null;
+    protected AwiConfig loadConfiguration() {
+        return ConfigUtils.readConfiguration();
     }
 }